Adiciona user_id para os todos

+++ todo-laravel/app/Http/Controllers/TodosController.php
@@ -18,7 +18,7 @@ class TodosController extends Controller
 
     public function index()
     {
-        return Todo::all();
+        return Todo::where('user_id', auth()->user()->id)->get();
     }
 
     /**
@@ -36,7 +36,11 @@ class TodosController extends Controller
             'completed' => 'required|boolean'
         ]);
 
-        $todo = Todo::create($data);
+        $todo = Todo::create([
+            'user_id' => auth()->user()->id,
+            'title' => $request->title,
+            'completed' => $request->completed,
+        ]);
 
         return response($todo, 201);
     }
@@ -63,6 +67,11 @@ class TodosController extends Controller
      */
     public function update(UpdateTodoRequest $request, Todo $todo)
     {
+
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $data = $request->validate([
             'title' => 'required|string',
             'completed' => 'required|boolean'
@@ -79,7 +88,7 @@ class TodosController extends Controller
             'completed' => 'required|boolean',
         ]);
 
-        Todo::query()->update($data);
+        Todo::where('user_id', auth()->user()->id)->update($data);
 
         return response('Updated', 200);
     }
@@ -92,6 +101,11 @@ class TodosController extends Controller
      */
     public function destroy(Todo $todo)
     {
+
+        if ($todo->user_id !== auth()->user()->id) {
+            return response()->json('Unauthorized', 401);
+        }
+
         $todo->delete();
 
         return response('Delted Todo Item', 200);
@@ -99,6 +113,21 @@ class TodosController extends Controller
 
     public function destroyCompleted(Request $request)
     {
+
+        $todosToDelete = $request->todos;
+
+        $userTodoIds = auth()->user()->todos->map(function ($todo) {
+            return $todo->id;
+        });
+
+        $valid = collect($todosToDelete)->every(function ($value, $key) use ($userTodoIds) {
+            return $userTodoIds->contains($value);
+        });
+
+        if (!$valid) {
+            return response()->json('Unauthorized', 401);
+        }
+        
 
         $request->validate([
             'todos' => 'required|array',
